small

+++ projs/022-digital-clock-raylib/src/libs/draw.c
@@ -14,26 +14,26 @@ int numbers_lookup_table[10][7] = {
   {true, true, true, true, true, false, true}, // 9
 };
 
-const Vector2 start_pos = {LENGTH+ OFFSET, W_HEIGHT / 2};
+const Vector2 start_pos = {LENGTH+ OFFSET, (float) W_HEIGHT / 2};
 
 void paint_triangle_strip(Vector2 center, bool vertical, Color color) {
     Vector2 a, b ,c ,d, e, f;
     int count = 6;
     
     if(vertical) {
-        a = (Vector2){center.x, center.y  - LENGTH / 2 - TIP};
-        b = (Vector2){center.x + THICKNESS / 2 , center.y - LENGTH / 2};
-        c = (Vector2){center.x - THICKNESS / 2 , center.y - LENGTH / 2};
-        d = (Vector2){center.x + THICKNESS / 2, center.y + LENGTH / 2};
-        e = (Vector2){center.x, center.y  + LENGTH / 2 + TIP };
-        f = (Vector2){center.x - THICKNESS / 2, center.y + LENGTH / 2};
+        a = (Vector2){center.x, center.y  - (float) LENGTH / 2 - TIP};
+        b = (Vector2){center.x + (float) THICKNESS / 2 , center.y - (float) LENGTH / 2};
+        c = (Vector2){center.x - (float) THICKNESS / 2 , center.y - (float) LENGTH / 2};
+        d = (Vector2){center.x + (float) THICKNESS / 2, center.y + (float) LENGTH / 2};
+        e = (Vector2){center.x, center.y  + (float) LENGTH / 2 + TIP };
+        f = (Vector2){center.x - (float) THICKNESS / 2, center.y + (float) LENGTH / 2};
     } else {
-        a = (Vector2){center.x - LENGTH / 2 - TIP, center.y};
-        b = (Vector2){center.x - LENGTH / 2 , center.y - THICKNESS / 2};
-        c = (Vector2){center.x - LENGTH / 2 , center.y + THICKNESS / 2};
-        d = (Vector2){center.x + LENGTH / 2, center.y - THICKNESS / 2};
-        e = (Vector2){center.x + LENGTH / 2 + TIP, center.y };
-        f = (Vector2){center.x + LENGTH / 2, center.y + THICKNESS / 2};
+        a = (Vector2){center.x - (float) LENGTH / 2 - TIP, center.y};
+        b = (Vector2){center.x - (float) LENGTH / 2 , center.y - (float) THICKNESS / 2};
+        c = (Vector2){center.x - (float) LENGTH / 2 , center.y + (float) THICKNESS / 2};
+        d = (Vector2){center.x + (float) LENGTH / 2, center.y - (float) THICKNESS / 2};
+        e = (Vector2){center.x + (float) LENGTH / 2 + TIP, center.y };
+        f = (Vector2){center.x + (float) LENGTH / 2, center.y + (float) THICKNESS / 2};
     }
    
     Vector2 points[]  = {a, c, b, f, d, e};
@@ -45,10 +45,10 @@ void draw_colon(Vector2 center) {
     
     bool visible = ((int)(t * 2) % 2) == 0;
     
-    Vector2 top = (Vector2){center.x, center.y - LENGTH / 2 - OFFSET / 2};
-    Vector2 bottom = (Vector2){center.x, center.y + LENGTH / 2 + OFFSET / 2};
-    DrawCircle(top.x, top.y, THICKNESS * 2/3, visible ? ON_COLOR : OFF_COLOR);
-    DrawCircle(bottom.x, bottom.y, THICKNESS * 2/3, visible ? ON_COLOR : OFF_COLOR);
+    Vector2 top = (Vector2){center.x, center.y - (float) LENGTH / 2 - (float) OFFSET / 2};
+    Vector2 bottom = (Vector2){center.x, center.y + (float) LENGTH / 2 + (float) OFFSET / 2};
+    DrawCircle(top.x, top.y, (float) THICKNESS * 2/3, visible ? ON_COLOR : OFF_COLOR);
+    DrawCircle(bottom.x, bottom.y, (float) THICKNESS * 2/3, visible ? ON_COLOR : OFF_COLOR);
 }
 
 
@@ -59,10 +59,10 @@ void draw_digit(int digit, Vector2 center) {
     Vector2 middle = (Vector2){center.x , center.y};
     Vector2 bottom = (Vector2){center.x , center.y + LENGTH + OFFSET};
     
-    Vector2 top_left = {center.x - LENGTH / 2  - OFFSET/2, center.y - LENGTH/2 - OFFSET /2}; 
-    Vector2 top_right = (Vector2){center.x + LENGTH / 2 + OFFSET/2, center.y - LENGTH/2 - OFFSET /2};
-    Vector2 bottom_right = (Vector2){center.x + LENGTH / 2 + OFFSET/2, center.y + LENGTH /2 + OFFSET /2};
-    Vector2 bottom_left = (Vector2){center.x - LENGTH / 2 - OFFSET/2, center.y + LENGTH / 2 + OFFSET / 2};
+    Vector2 top_left = {center.x - (float) LENGTH / 2  - (float) OFFSET/2, center.y - (float) LENGTH/2 - (float) OFFSET /2}; 
+    Vector2 top_right = (Vector2){center.x + (float) LENGTH / 2 + (float) OFFSET/2, center.y - (float) LENGTH/2 - (float) OFFSET /2};
+    Vector2 bottom_right = (Vector2){center.x + (float) LENGTH / 2 + (float) OFFSET/2, center.y + (float) LENGTH/2 + (float) OFFSET /2};
+    Vector2 bottom_left = (Vector2){center.x - (float) LENGTH / 2 - (float) OFFSET/2, center.y + (float) LENGTH / 2 + (float) OFFSET / 2};
    
     
     paint_triangle_strip(top_left, true, color_flags[0] ? ON_COLOR : OFF_COLOR);
